#304: Moved and renamed the file containing ResistancesElementaires
It no longer belongs to the Objet area.
